backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
nh3==0.3.2
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.3
webencodings==0.5.1
